Made the BasicCompletableFuture inner class of the ExecutionServiceImpl
